requirements.txt
fastapi
uvicorn[standard]
requests
pillow
python-dotenv
jinja2
colorthief
numpy
scipy